Random Seed: Add the ability to generate a random seed.
It is useful to be able to seed the RNG with a seed that is based on
the current time. This ensures different simulations generate
different results. However simulating with a known seed is also
useful.

This commit will use an input seed if that seed is >0. Otherwise it
will generate a seed based on the current system time. In all cases
the seed used is written to the output file.

Fixes #10.